Cast loans to invest to int

+++ utils.py
@@ -46,7 +46,7 @@ def get_loans(headers, loan_grade, min_probability_score, logger):
 
 
 def get_max_number_loans(available_cash, max_loan_invest):
-    return math.floor(available_cash / max_loan_invest)
+    return int(math.floor(available_cash / max_loan_invest))
 
 
 def header_builder(authorization_token):
